# docs/swagger.yaml
basePath: /
definitions:
...
  dtos.Character:
    properties:
      avatarUrl:
        type: string
      class:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      level:
        type: integer
      name:
        type: string
      playerId:
        type: string
      sheetUrl:
        type: string
      updatedAt:
        type: string
    type: object
  dtos.GameSystem:
    properties:
      badgeColor:
        type: string
      id:
        type: string
      isCurated:
        type: boolean
      name:
        type: string
      slug:
        type: string
    type: object
...
  internal_handlers.CreateCharacterRequest:
    properties:
      avatarUrl:
        type: string
      class:
        type: string
      description:
        type: string
      level:
        type: integer
      name:
        type: string
      sheetUrl:
        type: string
    type: object
  internal_handlers.CreateGameSystemRequest:
    properties:
      name:
        type: string
    type: object
...
  internal_handlers.EditCharacterRequest:
    properties:
      avatarUrl:
        type: string
      class:
        type: string
      description:
        type: string
      level:
        type: integer
      name:
        type: string
      sheetUrl:
        type: string
    type: object
...
  internal_handlers.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
host: localhost:5137
info:
  contact: {}
  description: Session, game system, campaign and character API for Session Zero
  title: Session Zero Session Service
  version: "1.0"
paths:
...
  /v1/characters:
    post:
      consumes:
      - application/json
      parameters:
      - description: Character data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateCharacterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.Character'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create a character
      tags:
      - characters
  /v1/characters/{id}:
    delete:
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Delete a character
      tags:
      - characters
    get:
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.Character'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get character by ID
      tags:
      - characters
    patch:
      consumes:
      - application/json
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.EditCharacterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.Character'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Edit a character
      tags:
      - characters
  /v1/characters/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.Character'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List my characters
      tags:
      - characters
  /v1/game-systems:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.GameSystem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Search game systems
      tags:
      - game-systems
    post:
      consumes:
      - application/json
      parameters:
      - description: Game system name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateGameSystemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.GameSystem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Add a custom game system
      tags:
      - game-systems
  /v1/game-systems/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.GameSystem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Get all game systems
      tags:
      - game-systems
  /v1/game-systems/curated:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.GameSystem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Get curated game systems
      tags:
      - game-systems
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"